// src/app/models/EducationInfoDto.ts
export class EducationInfo {
    egitim_durumu?: string;
    baslangicTarihi?: string;
    bitisTarihi?: string;
    devamEdiyorum?: boolean;
    terk?: boolean;
    diploma_not_sistemi?: string;
    diploma_notu?: string;
    lise?: string;
    lise_tipi?: string;
    lise_bolum?: string;
    universite?: string;
    fakulte?: string;
    bolum?: string;
    ogretim_tipi?: string;
    ogretim_dili?: string;
    burs_tipi?: string;
    is_yandal?: boolean;
    aciklama?: string;
}